Add files via upload
Updatedd cleint, performs the same tasks and does the same thing, except its more clean and the design has been improved. Changes like the connection boolean has been added and is called isConnected, also the message of server connection successful is displayed now with the host number. Added the ClientGUI but it is very very basic for now, performs the same exact functions as the client class does, buttons and the windows work as well. It is very messy at this stage, needs more design, planning on adding additonal features like the account name somewhere on the windwo plus some visuals like photos and logos.