Add files via upload
Reformed Client.java to fix login/signup flow, prevent null/EOF crashes, and ensure proper server-client communication using safe message handling and unqualified enum constants. Fully functional post-login menu now works reliably. Also worked on the StaffMember class and the list, will use it later, the client is modified enough for now, so next i will work on the separation between the staff and the user, the gui will be worked on after I am done with the client and the client is ready for implementation fro later of check in/check out functions + additional functions for location to search/add it but the location class has to be implemented first.