Refactored Item class and Added Location Specific Lists
Item class now has owner and lcoation attribute instead of qunatity and owners. item class doesnt save reserved yet so it would be nice if someone could add that.
Each location will generate text files for specific lists, i havent checked if adding a staff member or an item to a list works yet but if anything it should be just about there.