Add files via upload
I changed the server, the client and now the gui. One logic issue, I can add strikes but they are added to the staff member that is logged in. And I can reserve books but they’re again reserved by the staff. Gotta fix that. Also I changed Joshua’s item and item list so it’s serializable and changed server+client with that, so it cna work with the updated files that joshua made. Added one more header to easily make a non-staff member a staff member.